+++ pumpkin-inventory/src/crafting/crafting_screen_handler.rs
@@ -2,8 +2,7 @@ use std::sync::Arc;
 use std::sync::atomic::AtomicU8;
 
 use async_trait::async_trait;
-use pumpkin_data::recipes::{CraftingRecipeTypes, RECIPES_CRAFTING, RecipeResultStruct};
-use pumpkin_data::tag::Tagable;
+
 use pumpkin_world::inventory::Inventory;
 use pumpkin_world::item::ItemStack;
 use tokio::sync::Mutex;
@@ -13,10 +12,8 @@ use crate::screen_handler::{
 };
 use crate::slot::{NormalSlot, Slot};
 
-use super::recipes::{RecipeFinderScreenHandler, RecipeInputInventory};
+use super::recipes::RecipeInputInventory;
 
-// TODO: Implement ResultSlot
-// CraftingResultSlot.java
 #[derive(Debug)]
 pub struct ResultSlot {
     pub inventory: Arc<dyn RecipeInputInventory>,
@@ -33,171 +30,6 @@ impl ResultSlot {
             id: AtomicU8::new(0),
             result: Arc::new(Mutex::new(ItemStack::EMPTY)),
         }
-    }
-
-    async fn match_recipe(&self) -> Option<&RecipeResultStruct> {
-        let mut count: usize = 0;
-
-        for i in 0..self.inventory.size() {
-            let slot = self.inventory.get_stack(i).await;
-            let slot = slot.lock().await;
-            if !slot.is_empty() {
-                count += 1;
-            }
-        }
-
-        'next_recipe: for recipe in RECIPES_CRAFTING {
-            match recipe {
-                CraftingRecipeTypes::CraftingShaped {
-                    key,
-                    pattern,
-                    result,
-                    ..
-                } => {
-                    if pattern.len() > self.inventory.get_height()
-                        || pattern.first().unwrap().len() > self.inventory.get_width()
-                    {
-                        continue;
-                    }
-
-                    if count
-                        != pattern
-                            .iter()
-                            .map(|l| l.chars().filter(|c| *c != ' ').count())
-                            .sum::<usize>()
-                    {
-                        continue;
-                    }
-
-                    for y_offset in 0..=(self.inventory.get_height() - pattern.len()) {
-                        'next_offset: for x_offset in
-                            0..=(self.inventory.get_width() - pattern[0].len())
-                        {
-                            // Check if pattern matches
-                            for y in 0..pattern.len() {
-                                for x in 0..pattern[y].len() {
-                                    let current_key = pattern[y].chars().nth(x).unwrap();
-                                    if current_key == ' ' {
-                                        continue;
-                                    }
-
-                                    let ingredient = key
-                                        .iter()
-                                        .find_map(|(k, v)| (*k == current_key).then_some(v))
-                                        .expect("Crafting recipe used invalid key");
-
-                                    let slot = self
-                                        .inventory
-                                        .get_stack(
-                                            (y + y_offset) * self.inventory.get_height()
-                                                + (x + x_offset),
-                                        )
-                                        .await;
-                                    let slot = slot.lock().await;
-
-                                    if !ingredient.match_item(slot.item) {
-                                        continue 'next_offset;
-                                    }
-                                }
-                            }
-
-                            // TODO: Apply components
-                            return Some(result);
-                        }
-                    }
-
-                    continue 'next_recipe;
-                }
-                CraftingRecipeTypes::CraftingShapeless {
-                    ingredients,
-                    result,
-                    ..
-                } => {
-                    if count != ingredients.len() {
-                        continue;
-                    }
-
-                    let mut ingredient_used = vec![false; ingredients.len()];
-                    'next_slot: for i in 0..self.inventory.size() {
-                        let slot = self.inventory.get_stack(i).await;
-                        let slot = slot.lock().await;
-
-                        if slot.is_empty() {
-                            continue 'next_slot;
-                        }
-
-                        for i in 0..ingredients.len() {
-                            if !ingredient_used[i] && ingredients[i].match_item(slot.item) {
-                                ingredient_used[i] = true;
-                                continue 'next_slot;
-                            }
-                        }
-
-                        continue 'next_recipe;
-                    }
-
-                    // TODO: Apply components
-                    return Some(result);
-                }
-                CraftingRecipeTypes::CraftingTransmute {
-                    input,
-                    material,
-                    result,
-                    ..
-                } => {
-                    if count != 2 {
-                        continue;
-                    }
-
-                    'item_stack: for i in 0..self.inventory.size() {
-                        let slot = self.inventory.get_stack(i).await;
-                        let slot = slot.lock().await;
-
-                        if slot.is_empty() {
-                            continue 'item_stack;
-                        }
-
-                        if !material.match_item(slot.item) && !input.match_item(slot.item) {
-                            continue 'next_recipe;
-                        }
-                    }
-
-                    // TODO: Copy components
-                    return Some(result);
-                }
-                CraftingRecipeTypes::CraftingDecoratedPot { .. } => {
-                    if count != 4
-                        || self.inventory.get_width() != 3
-                        || self.inventory.get_height() != 3
-                    {
-                        continue 'next_recipe;
-                    }
-
-                    for position in (1..=7).step_by(2) {
-                        let slot = self.inventory.get_stack(position).await;
-                        let slot = slot.lock().await;
-
-                        if slot.is_empty()
-                            || !slot
-                                .item
-                                .is_tagged_with("#minecraft:decorated_pot_ingredients")
-                                .unwrap()
-                        {
-                            continue 'next_recipe;
-                        }
-                    }
-
-                    // TODO: Handle side textures
-                    return Some(&RecipeResultStruct {
-                        id: "minecraft:decorated_pot",
-                        count: 1,
-                    });
-                }
-                CraftingRecipeTypes::CraftingSpecial => continue,
-            }
-        }
-
-        None
     }
 }
 
@@ -230,7 +62,7 @@ impl Slot for ResultSlot {
         // }
 
         // self.amount.store(0, Ordering::Relaxed);
-        // TODO: Unlock recipes
+        // TODO: Unlock recipes for the recipe book
 
         for i in 0..self.inventory.size() {
             let slot = self.inventory.get_stack(i).await;
@@ -300,6 +132,7 @@ impl ScreenHandlerListener for ResultSlot {
             .contains(&(slot as usize))
         {
             let result = self
+                .inventory
                 .match_recipe()
                 .await
                 .map(ItemStack::from)
@@ -316,11 +149,8 @@ impl ScreenHandlerListener for ResultSlot {
     }
 }
 
-// AbstractCraftingScreenHandler.java
 #[async_trait]
-pub trait CraftingScreenHandler<I: RecipeInputInventory>:
-    RecipeFinderScreenHandler + ScreenHandler
-{
+pub trait CraftingScreenHandler<I: RecipeInputInventory>: ScreenHandler {
     async fn add_recipe_slots(&mut self, crafing_inventory: Arc<dyn RecipeInputInventory>) {
         let result_slot = Arc::new(ResultSlot::new(crafing_inventory.clone()));
         self.add_slot(result_slot.clone());
