+++ pumpkin-inventory/src/crafting/recipes.rs
@@ -1,18 +1,12 @@
+use async_trait::async_trait;
+
+use pumpkin_data::recipes::RecipeResultStruct;
 use pumpkin_world::inventory::Inventory;
 
-// RecipeMatcher.java
-pub struct RecipeMatcher {}
-
-// RecipeFinder.java
-pub struct RecipeFinder {}
-
-// AbstractRecipeScreenHandle.java
-pub trait RecipeFinderScreenHandler {}
-
+#[async_trait]
 pub trait RecipeInputInventory: Inventory {
     fn get_width(&self) -> usize;
     fn get_height(&self) -> usize;
-    //fn get_held_stacks(), Get a lock on the inventory instead
-    // createRecipeInput
-    // createPositionedRecipeInput
+
+    async fn match_recipe(&self) -> Option<&RecipeResultStruct>;
 }
